rename ipc module
