+++ app/actions/video.js
@@ -4,7 +4,7 @@ import {
    REQUEST_FILE,
    RECEIVE_FILE,
 } from '../services/signals';
-import { requestFile, reportError } from '../services/ipcDispatcher';
+import { requestFile, reportError } from '../services/ipc';
 import { ADD_VIDEO_REQUEST, ADD_VIDEO_SUCCESS } from './videos';
 
 export const UPLOAD_VIDEO_REQUEST = 'UPLOAD_VIDEO_REQUEST';
